Fix DeploymentController Progressing
The Progression condition should be set only when the Deployment has a new
content and is actively rolling out its update. After all replicas are
updated, the Progressing condition should be false, even when some pods are
missing. E.g. because a node is drained, something evicted them or so on.

Use Deployment condition Progressing with reason NewReplicaSetAvailable to
detect that Deployment has been fully rolled out in the past.